# ct313hm02-project-DrStone113/backend-api/docs/openapiSpec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fashivo - Fashion E-commerce Website API",
    "version": "1.0.0",
    "description": "API for Fashivo fashion e-commerce website",
    "contact": {
      "name": "Huỳnh Thanh Nhuận, Thạch Nguyễn Khang",
      "email": "khangb2205987@student.ctu.edu.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/carts": {
      "get": {
        "summary": "Get all carts",
        "description": "Retrieve a list of all shopping carts with pagination and optional user ID filter.",
        "operationId": "getAllCarts",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "description": "Filter carts by user ID."
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of carts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "carts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CartWithItems"
                          }
                        },
                        "metadata": {
                          "$ref": "#/components/schemas/PaginationMetadata"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      },
      "post": {
        "summary": "Create or update a shopping cart",
        "description": "Adds items to a user's cart. If the cart doesn't exist, it creates one. If the product already exists in the cart, it updates the quantity. This operation expects multipart/form-data where 'items' is a JSON string.",
        "operationId": "createCart",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "format": "int32",
                    "description": "ID of the user for whom the cart is being created/updated.",
                    "minimum": 1
                  },
                  "items": {
                    "type": "string",
                    "description": "JSON string representing an array of cart items. Example: `[{\"product_id\": 1, \"quantity\": 2}, {\"product_id\": 3, \"quantity\": 1}]`",
                    "example": "[{\"product_id\": 1, \"quantity\": 2}, {\"product_id\": 3, \"quantity\": 1}]"
                  }
                },
                "required": ["user_id", "items"]
              },
              "encoding": {
                "items": {
                  "contentType": "application/json"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cart created or updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": {
                          "$ref": "#/components/schemas/CartWithItems"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      },
      "delete": {
        "summary": "Delete all shopping carts",
        "description": "Deletes all shopping carts and their items. (Use with caution)",
        "operationId": "deleteAllCarts",
        "responses": {
          "204": {
            "description": "All carts deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      }
    },
    "/api/v1/carts/user/{userId}": {
      "get": {
        "summary": "Get a user's shopping cart",
        "description": "Retrieve a specific user's shopping cart, including all items within it.",
        "operationId": "getCartByUserId",
        "parameters": [
          {
            "$ref": "#/components/parameters/userIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User's cart data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": {
                          "$ref": "#/components/schemas/CartWithItems"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      }
    },
    "/api/v1/carts/{id}": {
      "get": {
        "summary": "Get a shopping cart by ID",
        "description": "Retrieve a specific shopping cart by its ID, including all items within it.",
        "operationId": "getCartById",
        "parameters": [
          {
            "$ref": "#/components/parameters/cartIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cart": {
                          "$ref": "#/components/schemas/CartWithItems"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      },
      "delete": {
        "summary": "Delete a shopping cart by ID",
        "description": "Deletes a shopping cart and all its associated items.",
        "operationId": "deleteCart",
        "parameters": [
          {
            "$ref": "#/components/parameters/cartIdParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Cart deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      }
    },
    "/api/v1/carts/{id}/item/{productId}": {
      "put": {
        "summary": "Update quantity of a specific item in a cart",
        "description": "Updates the quantity of a product within a specified shopping cart. This operation expects multipart/form-data for 'quantity'.",
        "operationId": "updateCartItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the cart",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "ID of the product within the cart",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "description": "New quantity for the product in the cart",
                    "minimum": 1
                  }
                },
                "required": ["quantity"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart item quantity updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cartItem": {
                          "$ref": "#/components/schemas/CartItemDetailed"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      },
      "delete": {
        "summary": "Delete a specific item from a cart",
        "description": "Removes a product from a specified shopping cart.",
        "operationId": "deleteCartItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the cart",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "ID of the product to remove from the cart",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cart item deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        },
        "tags": ["Cart"]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CartItemDetailed": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32", "example": 1 },
          "cart_id": { "type": "integer", "format": "int32", "example": 1 },
          "product_id": { "type": "integer", "format": "int32", "example": 1 },
          "quantity": { "type": "integer", "example": 2 },
          "item_price": { "type": "number", "format": "float", "example": 699.99, "description": "Price of the product when it was added to the cart" },
          "product_name": { "type": "string", "example": "Smartphone" },
          "product_description": { "type": "string", "example": "Latest model smartphone" },
          "current_product_price": { "type": "number", "format": "float", "example": 729.99, "description": "Current price of the product" },
          "product_image_url": { "type": "string", "example": "/public/img/products/product-12345.jpg" },
          "product_available": { "type": "boolean", "example": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32", "example": 1 },
          "user_id": { "type": "integer", "format": "int32", "example": 1 },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CartWithItems": {
        "type": "object",
        "allOf": [
          { "$ref": "#/components/schemas/Cart" },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CartItemDetailed"
                }
              }
            }
          }
        ]
      },
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "totalRecords": { "type": "integer", "example": 50 },
          "currentPage": { "type": "integer", "example": 1 },
          "totalPages": { "type": "integer", "example": 5 },
          "firstPage": { "type": "integer", "example": 1 },
          "lastPage": { "type": "integer", "example": 5 },
          "limit": { "type": "integer", "example": 10 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "error" },
          "message": { "type": "string", "example": "Something went wrong" },
          "statusCode": { "type": "integer", "example": 500 }
        }
      }
    },
    "parameters": {
...
      "limitParam": {
        "name": "limit",
        "description": "Number of records per page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        },
        "required": false
      },
      "pageParam": {
        "name": "page",
        "description": "Page number of records",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "required": false
      },
...
      "userIdParam": {
        "name": "userId",
        "description": "User ID",
        "required": true,
        "in": "path",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1
        }
      },
      "cartIdParam": {
        "name": "id",
        "description": "Cart ID",
        "required": true,
        "in": "path",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
...
}